🎨 알림기능 수정중 (#115)
* :art: 알림기능 수정중
