+++ alarm/consumers.py
@@ -1,24 +1,9 @@
 import json
-from asgiref.sync import async_to_sync
 from channels.db import database_sync_to_async
-from channels.layers import get_channel_layer
 from channels.generic.websocket import AsyncWebsocketConsumer
 
 from .models import Alarm
 from .serializers import AlarmSerializer
-from django.db.models.signals import post_save
-from django.dispatch import receiver
-
-
-@receiver(post_save, sender=Alarm)
-def send_alarm(sender, instance, created, **kwargs):
-    if created:
-        channel_layer = get_channel_layer()
-        user_id = instance.user.id
-        async_to_sync(channel_layer.group_send)(
-            f"user{user_id}",
-            {"type": "send_alarm", "message": AlarmSerializer(instance).data},
-        )
 
 
 def get_alarm(user):
